% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_Regmode.R
\name{assign_Regmode}
\alias{assign_Regmode}
\title{This function reads the results of the function getFCs and assigns a
regulatory mode to each gene based on the Fold Change (FC) for RNA-Seq Counts,
Ribo-Seq Counts, or TE. It is intended for internal use and its output is
the output of get_FCs.R}
\usage{
assign_Regmode(res_df)
}
\arguments{
\item{res_df}{A dataframe, output of the function getFCs}
}
\value{
A dataframe, with two extra columns with info on the Regulatory Mode
}
\description{
This function reads the results of the function getFCs and assigns a
regulatory mode to each gene based on the Fold Change (FC) for RNA-Seq Counts,
Ribo-Seq Counts, or TE. It is intended for internal use and its output is
the output of get_FCs.R
}
\examples{
# For testing purposes of this internal function a mock-file is provided
# within the extdata folder.
mockdata <- system.file("extdata", "mock_get_FCs_result.tsv",
package = "terapadog")
result <- assign_Regmode(mockdata)
# Only the head of the result file will be returned
print(head(result))
}
\keyword{internal}
