% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_helpers.R
\name{check_input_df}
\alias{check_input_df}
\title{Check if a data frame is not empty (no rows, no columns, or NULL)}
\usage{
check_input_df(df1)
}
\arguments{
\item{df1}{A dataframe.}
}
\value{
NULL. Throws an error if df is empty or NULL.
}
\description{
Check if a data frame is not empty (no rows, no columns, or NULL)
}
\examples{
#' # Internal function, code cannot be run from here.
\dontrun{
rna_df <- data.frame(Sample1 = c(1,2,3), Sample2 = c(1,2,3), Sample3 = c(1,2,3))
check_input_df(rna_df)
}
}
\keyword{internal}
