% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_helpers.R
\name{check_integer_values}
\alias{check_integer_values}
\title{Check and Convert Dataframe Columns to Integers}
\usage{
check_integer_values(df)
}
\arguments{
\item{df}{A dataframe to be checked and potentially modified.}
}
\value{
A dataframe with numeric columns as integers.
}
\description{
This function checks whether each numeric column in the provided dataframe
contains only integer values. If a column contains floating-point numbers,
it rounds the values and converts them to integers.
}
\keyword{internal}
