% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_helpers.R
\name{check_matching_colnames}
\alias{check_matching_colnames}
\title{Check if two data frames have the same column names}
\usage{
check_matching_colnames(df1, df2)
}
\arguments{
\item{df1}{A dataframe.}

\item{df2}{A dataframe.}
}
\value{
NULL. Throws an error if column names do not match.
}
\description{
This function verifies that two data frames (RIBO and RNA) contain
the same set of column names, regardless of their order.
This is important for how terapadog handles these matching samples during
shuffles.
}
\keyword{internal}
