% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_helpers.R
\name{detect_separator}
\alias{detect_separator}
\title{Detect the separator used in a file}
\usage{
detect_separator(path)
}
\arguments{
\item{path}{A string representing the file path.}
}
\value{
A string representing the separator (comma or tab).
}
\description{
This function determines whether the file uses a comma or tab as a separator
based on the file extension (.csv or .tsv).
}
\keyword{internal}
