% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_FCs.R
\name{get_FCs}
\alias{get_FCs}
\title{This function execute the Differential Translation Analysis on its own
using DeltaTE.
The output is a dataframe with the FC in mRNA counts, RIBO counts or TE
between the conditions in exam.}
\usage{
get_FCs(expression.data, exp_de, paired = FALSE)
}
\arguments{
\item{expression.data}{A matrix containing the counts from RNA and RIBO
samples.}

\item{exp_de}{A dataframe containing information regarding the samples.
It has number of rows equal to the columns of esetm.}

\item{paired}{Logical. Default is false. Set to TRUE if the experiment
has paired samples in its design.}
}
\description{
This function execute the Differential Translation Analysis on its own
using DeltaTE.
The output is a dataframe with the FC in mRNA counts, RIBO counts or TE
between the conditions in exam.
}
