% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_results_viz.R
\name{get_gene_from_kegg_pathway}
\alias{get_gene_from_kegg_pathway}
\title{This function retrieves, for a given pathway id, the member genes
GeneID (ensembl) and gene name.
It then extracts their rows from the results produced by get_FC, to allow
the manual inspection of the genes within a pathway of interest.}
\usage{
get_gene_from_kegg_pathway(
  fc_results,
  pathway_id,
  organism_prefix = "hsa",
  ensembl_dataset = "hsapiens_gene_ensembl"
)
}
\arguments{
\item{fc_results}{A dataframe containing the results of the get_FC() function.}

\item{pathway_id}{The kegg ID for the pathway of interest, as reported in the
terapadog() results "ID" field.}

\item{organism_prefix}{Character. The three-letter code for the organism
of interest. default is H. sapiens ("hsa").}

\item{ensembl_dataset}{Character. Defines the ensembl dataset to use. Default
is "hsapiens_gene_ensembl".}
}
\value{
A dataframe with the results of a Differential Translation Analysis
for only the genes belonging to the specified pathway.
The Ensembl GeneID and the gene Symbol are also provided.
}
\description{
This function retrieves, for a given pathway id, the member genes
GeneID (ensembl) and gene name.
It then extracts their rows from the results produced by get_FC, to allow
the manual inspection of the genes within a pathway of interest.
}
\examples{
\dontrun{
# DO NOT RUN: 1 - Queries KEGG and Ensembl - requires network access.
# 2 - KEGG does not allow to redistribute their data. Only research purposes.
# As such, this example cannot be run to build documentation.
fc_results_path <- system.file("extdata", "fc_results_example.csv",
package = "terapadog")
fc_results <- read.table(fc_results_path, sep = ",", header = TRUE)
mapped <- get_gene_from_kegg_pathway(fc_results, pathway_id = "4350")
}
}
