% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_converter.R
\name{id_converter}
\alias{id_converter}
\title{Convert the human gene identifier (hgnc_symbol or ensembl_gene_id) to
entrezgene_id format for the analysis.}
\usage{
id_converter(esetm, id_type)
}
\arguments{
\item{esetm}{A dataframe with the gene count values.}

\item{id_type}{A string representing the type of ID given as input. Must be
either hgnc_symbol or ensembl_gene_id.}
}
\value{
A dataframe with gene IDs in the entrezgene_id format.
}
\description{
Convert the human gene identifier (hgnc_symbol or ensembl_gene_id) to
entrezgene_id format for the analysis.
}
\examples{
# To showcase this internal function, a small example is made.
gene_ids <- c('ENSG00000103197', 'ENSG00000008710', 'ENSG00000167964')
esetm <- data.frame(
Sample1 = c(2.5, 3.1, 5.2),
Sample2 = c(4.1, 2.9, 6.3),
Sample3 = c(1.5, 3.7, 4.8),
row.names = gene_ids)
# Call the function
esetm <- id_converter(esetm, "ensembl_gene_id")
print(head(esetm))
}
