% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_results_viz.R
\name{plot_pathway_changes}
\alias{plot_pathway_changes}
\title{This function plots, for a given pathway id, the member genes RNA, RIBO, and
TE fold changes on the relative pathway (3 plots).
Only significant genes are plotted.
Plots are saved as .png files.}
\usage{
plot_pathway_changes(
  fc_results,
  pathway_id,
  alpha_threshold = 0.05,
  organism_prefix = "hsa"
)
}
\arguments{
\item{fc_results}{A dataframe containing the results of the get_FC() function.}

\item{pathway_id}{The kegg ID for the pathway of interest, as reported in the
terapadog() results "ID" field.}

\item{alpha_threshold}{float. Defines the threshold for alpha. Values above
are considered non-significant.}

\item{organism_prefix}{Character. The three-letter code for the organism
of interest. default is H. sapiens ("hsa").}
}
\value{
Saves three pictures in the working directory. Behaviour inherited
by pathview. Changing the workdir changes the output directory.
}
\description{
This function plots, for a given pathway id, the member genes RNA, RIBO, and
TE fold changes on the relative pathway (3 plots).
Only significant genes are plotted.
Plots are saved as .png files.
}
\examples{
\dontrun{
# DO NOT RUN: 1 - Queries KEGG - requires network access.
# 2 - KEGG does not allow to redistribute their data in vignettes or docs.
# Only research purposes.
fc_results_path <- system.file("extdata", "fc_results_example.csv",
package = "terapadog")
fc_results <- read.table(fc_results_path, sep = ",", header = TRUE)
plot_pathway_changes(fc_results, "4350")
}
}
