% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_Regmode.R
\name{assign_Regmode}
\alias{assign_Regmode}
\title{This function reads the results of the function getFCs and assigns a
regulatory mode to each gene based on the Fold Change (FC) for RNA-Seq Counts,
Ribo-Seq Counts, or TE.}
\usage{
assign_Regmode(res_df)
}
\arguments{
\item{res_df}{A dataframe, output of the function getFCs}
}
\value{
A dataframe, with two extra columns with info on the Regulatory Mode
}
\description{
This function reads the results of the function getFCs and assigns a
regulatory mode to each gene based on the Fold Change (FC) for RNA-Seq Counts,
Ribo-Seq Counts, or TE.
}
\keyword{internal}
