% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_converter.R
\name{id_converter}
\alias{id_converter}
\title{Convert the gene identifier to entrezgene_id format for the analysis.
Works on human only!}
\usage{
id_converter(esetm, id_type)
}
\arguments{
\item{esetm}{A dataframe with the gene count values.}

\item{id_type}{A string representing the type of ID given as input. Must be
either hgnc_symbol or ensembl_gene_id.}
}
\value{
A dataframe with gene IDs in the entrezgene_id format.
}
\description{
Convert the gene identifier to entrezgene_id format for the analysis.
Works on human only!
}
\keyword{internal}
