% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDTA.R
\name{plotDTA}
\alias{plotDTA}
\title{This function will plot an interactive html plot of the results of get_FCs.R
That is to say, a plot of the genes undergoing translational regulation,
coloured by RegMode.
Genes whose RegMode was Undeterminable or Undetermined are omitted.}
\usage{
plotDTA(FC_results, path)
}
\arguments{
\item{FC_results}{A matrix containing the counts from RNA and RIBO
samples.}

\item{path}{A string, pointing to where to save the html plot.}
}
\description{
This function will plot an interactive html plot of the results of get_FCs.R
That is to say, a plot of the genes undergoing translational regulation,
coloured by RegMode.
Genes whose RegMode was Undeterminable or Undetermined are omitted.
}
